Renamed all files and updated tests to match
